[ESP32] Refactor ESP32 all-clusters-app example (#15624)
* esp: refactor main.cpp in all-cluster-app

* move m5stack code to separate file

* esp: refactor button driver in all-cluster-app

* change M5Stack button driver from poll to isr
* delete unused head file in main.cpp

* esp: optimize all-cluster-app user task execution process

* use App Task to manage user tasks
* update copyright to latest
* optimize button debouncing mechanism
