+++ examples/all-clusters-app/esp32/main/include/AppEvent.h
@@ -16,10 +16,36 @@
  *    limitations under the License.
  */
 
-#include "Globals.h"
+#pragma once
+#include "esp_system.h"
 
-LEDWidget statusLED1;
-LEDWidget statusLED2;
-BluetoothWidget bluetoothLED;
-WiFiWidget wifiLED;
-WiFiWidget pairingWindowLED;
+struct AppEvent;
+typedef void (*EventHandler)(AppEvent *);
+
+struct AppEvent
+{
+    enum AppEventTypes
+    {
+        kEventType_Button = 0,
+        kEventType_Timer,
+        kEventType_Light,
+        kEventType_Install,
+    };
+
+    uint16_t mType;
+
+    union
+    {
+        struct
+        {
+            uint8_t mPinNo;
+            uint8_t mAction;
+        } mButtonEvent;
+        struct
+        {
+            void * mContext;
+        } mTimerEvent;
+    };
+
+    EventHandler mHandler;
+};
