+++ examples/all-clusters-app/esp32/main/include/Globals.h
@@ -1,6 +1,7 @@
 /*
  *
  *    Copyright (c) 2020 Project CHIP Authors
+ *    All rights reserved.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -27,3 +28,4 @@ extern LEDWidget statusLED2;
 extern BluetoothWidget bluetoothLED;
 extern WiFiWidget wifiLED;
 extern const chip::NodeId kLocalNodeId;
+extern WiFiWidget pairingWindowLED;
