+++ examples/all-clusters-app/esp32/sdkconfig.defaults
