+++ examples/all-clusters-app/esp32/sdkconfig_m5stack.defaults
@@ -58,3 +58,9 @@ CONFIG_LWIP_HOOK_ND6_GET_GW_DEFAULT=y
 # Serial Flasher config
 CONFIG_ESPTOOLPY_FLASHSIZE_4MB=y
 CONFIG_ESPTOOLPY_FLASHSIZE="4MB"
+
+#disable Bluetooth modem sleep
+#enable it may cause GPIO ISR triggers continuously
+CONFIG_BTDM_CTRL_MODEM_SLEEP=n
+CONFIG_BTDM_CTRL_MODEM_SLEEP_MODE_ORIG=n
+CONFIG_BTDM_CTRL_LPCLK_SEL_MAIN_XTAL=n
