+++ examples/all-clusters-app/esp32/sdkconfig_m5stack_rpc.defaults
@@ -61,3 +61,10 @@ CONFIG_ENABLE_CHIP_SHELL=n
 # Serial Flasher config
 CONFIG_ESPTOOLPY_FLASHSIZE_4MB=y
 CONFIG_ESPTOOLPY_FLASHSIZE="4MB"
+
+
+#disable Bluetooth modem sleep
+#enable it may cause GPIO ISR triggers continuously
+CONFIG_BTDM_CTRL_MODEM_SLEEP=n
+CONFIG_BTDM_CTRL_MODEM_SLEEP_MODE_ORIG=n
+CONFIG_BTDM_CTRL_LPCLK_SEL_MAIN_XTAL=n
